+++ src/main/java/com/_ithon/speeksee/global/auth/service/GoogleOAuth2Client.java
@@ -1,4 +1,4 @@
-package com._ithon.speeksee.global.auth.oauth.service;
+package com._ithon.speeksee.global.auth.service;
 
 import java.util.Map;
